Made the u32 parse use the Parse trait
