+++ src/parse.rs
@@ -1,35 +1,10 @@
 use xmltree::Element;
-use failure::ResultExt;
-use ElementExt;
 
 use error::*;
 
 macro_rules! try {
     ($e:expr) => {
         $e.expect(concat!(file!(), ":", line!(), " ", stringify!($e)))
-    }
-}
-
-// TODO: Should work on &str not Element
-// TODO: `parse::u32` should not hide it's errors, see `BitRange::parse`
-pub fn u32(tree: &Element) -> Result<u32, SVDError> {
-    let text = tree.get_text()?;
-
-    if text.starts_with("0x") || text.starts_with("0X") {
-        u32::from_str_radix(&text["0x".len()..], 16).context(SVDErrorKind::Other(format!("{} invalid", text))).map_err(|e| e.into())
-    } else if text.starts_with('#') {
-        // Handle strings in the binary form of:
-        // #01101x1
-        // along with don't care character x (replaced with 0)
-        u32::from_str_radix(&str::replace(&text.to_lowercase()["#".len()..], "x", "0"), 2).context(SVDErrorKind::Other(format!("{} invalid", text))).map_err(|e| e.into())
-    } else if text.starts_with("0b"){
-        // Handle strings in the binary form of:
-        // 0b01101x1
-        // along with don't care character x (replaced with 0)
-        u32::from_str_radix(&str::replace(&text["0b".len()..], "x", "0"), 2).context(SVDErrorKind::Other(format!("{} invalid", text))).map_err(|e| e.into())
-
-    } else {
-        text.parse::<u32>().context(SVDErrorKind::Other(format!("{} invalid", text))).map_err(|e| e.into())
     }
 }
